+++ src/picasapy/app/i18n/picasapy_hu.ts
@@ -3860,6 +3860,64 @@ Egy gyűjtemény megnyitásához kattintson duplán a nevére, vagy kattintson a
     </message>
 </context>
 <context>
+    <name>PublishPanel</name>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Selection and Settings</source>
+        <translation>Kijelölés és beállítások</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Photo Size</source>
+        <translation>Fotóméret</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Name the Gift CD</source>
+        <translation>Az ajándék CD elnevezése</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>CD Name</source>
+        <translation>CD neve</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Limit 16 Characters</source>
+        <translation>Legfeljebb 16 karakter</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Erase Media</source>
+        <translation>Adathordozó törlése</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Include Picasa</source>
+        <translation>A Picasával együtt</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Backup Set</source>
+        <translation>Mentési készlet</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Size:</source>
+        <translation>Méret:</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Visibility:</source>
+        <translation>Láthatóság:</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/PublishPanel.qml" line="0"/>
+        <source>Sync:</source>
+        <translation>Szinkronizálás:</translation>
+    </message>
+</context>
+<context>
     <name>LightboxHeader</name>
     <message>
         <location filename="../qml/PicasaPy/LightboxHeader.qml" line="0"/>